+++ testsuite/test_tupleparse.py
@@ -22,14 +22,18 @@ class TupleTest(unittest.TestCase):
     """ Top level class for testing tupleparse"""
 
     def _run_single(self, obj):
-
+        """
+        Execute the test on a single object.  This test converts the object to
+        xml and then parses it.  The result must be the same as the original
+        object.
+        """
         # Convert object to xml
 
         xml = obj.tocimxml().toxml()
 
         # Parse back to an object
 
-        result = tupleparse.parse_any(tupletree.xml_to_tupletree(xml))
+        result = tupleparse.parse_any(tupletree.xml_to_tupletree_sax(xml))
 
         # Assert that the before and after objects should be equal
 
@@ -42,10 +46,13 @@ class RawXMLTest(unittest.TestCase):
     """Test parse of any XML to object"""
 
     def _run_single(self, xml, obj):
-
+        """
+        Parse the xml parameter to an object and compare it with the obj
+        parameter
+        """
         # Parse raw XML to an object
 
-        result = tupleparse.parse_any(tupletree.xml_to_tupletree(xml))
+        result = tupleparse.parse_any(tupletree.xml_to_tupletree_sax(xml))
 
         # Assert XML parses to particular Python object
 
@@ -57,7 +64,7 @@ class ParseCIMInstanceName(TupleTest):
     """Test parsing of CIMInstanceName objects."""
 
     def test_all(self):
-
+        """Test multiple instance name constructions"""
         self._run_single(CIMInstanceName('CIM_Foo'))
 
         self._run_single(CIMInstanceName('CIM_Foo',
@@ -84,7 +91,7 @@ class ParseCIMInstance(TupleTest):
     """Test parsing of CIMInstance objects."""
 
     def test_all(self):
-
+        """Test multiple CIMInstanceObjects"""
         self._run_single(CIMInstance('CIM_Foo'))
 
         self._run_single(
@@ -107,7 +114,7 @@ class ParseCIMClass(TupleTest):
     """Test parsing of CIMClass objects."""
 
     def test_all(self):
-
+        """test CIMClass Constructors"""
         self._run_single(CIMClass('CIM_Foo'))
         self._run_single(CIMClass('CIM_Foo', superclass='CIM_bar'))