+++ pywbem/_server.py
@@ -34,7 +34,8 @@ This chapter has the following sections:
 
 .. note::
 
-   At this point, the WBEM server API is experimental.
+   The WBEM server API has been introduced in v0.9.0 as experimental and
+   has been declared final in 0.10.0.
 
 .. _`Server Example`: