+++ docs/changes.rst
@@ -51,6 +51,9 @@ Enhancements
 * Declared the WBEM indications API and the WBEM server API to be final. These
   APIs had been introduced in v0.9.0 as experimental.
 
+* Added `git` as an OS-level dependency for development (it is used by GitPython
+  when building the documentation).
+
 
 Bug fixes
 ^^^^^^^^^